src/docker_multiagentes/linear_regression/requirements.txt
pandas
numpy
seaborn
scikit-learn
sqlalchemy
psycopg2-binary
